Introduce a "run mode"
We use the autoscaler to turn off certain deployments over the weekend. It seems wasteful to keep a process in our cluster that constantly consumes 100MB of memory just for doing its work on Fridays. So, we have set up the autoscaler as a CronJob, which required some minor changes to the startup code.